+++ .gitignore
@@ -5,3 +5,6 @@
 .vscode
 
 .DS_STORE
+
+build/
+original/
