+++ CMakeLists.txt
@@ -1,0 +1,21 @@
+cmake_minimum_required(VERSION 3.20.0)
+
+project(freecplus)
+
+set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/bin)
+set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/lib)
+set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/lib)
+
+set(CMAKE_C_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
+
+ADD_DEFINITIONS(
+-O3 -g -W -Wall
+ -Wunused-variable -Wunused-parameter -Wunused-function -Wunused
+ -Wno-deprecated -Woverloaded-virtual -Wwrite-strings
+ -D__WUR= -D_REENTRANT -D_FILE_OFFSET_BITS=64 -DTIXML_USE_STL
+)
+
+add_subdirectory(freecplus)
+add_subdirectory(demo)
+add_subdirectory(db)
