--- _cmpublic.h
@@ -1,59 +1,0 @@
-/****************************************************************************************/
-/*   程序名：_cmpublic.h，此程序是freecplus框架公用头文件，包含了应用开发常用的头文件。 */
-/*   作者：C语言技术网(www.freecplus.net) 日期：20190525。                              */
-/*                                                                                      */
-/*   freecplus开源许可协议正文如下：                                                    */
-/*     版权所有 (c) 2008-2020，码农有道，C语言技术网（www.freecplus.net）。             */
-/*     特此向任何得到本软件副本或相关文档的人授权：被授权人有权使用、复制、修改、       */
-/*   合并、出版、发布、散布、再授权和/或贩售软件及软件的副本，及授予被供应人同等        */
-/*   权利，只需服从以下义务：                                                           */
-/*     在软件和软件的所有副本中都必须包含以上版权声明和本许可声明。                     */
-/*     如果违返，无论在本软件的使用上或其他买卖交易中是否涉及合同，都属于侵权行为，     */
-/*   我们保留追究的权利。                                                               */
-/*   freecplus开源许可协议正文结束。                                                    */
-/****************************************************************************************/
-
-#ifndef _cmpublic_H
-#define _cmpublic_H
-
-#include <stdio.h>
-#include <utime.h>
-#include <string.h>
-#include <strings.h>
-#include <ctype.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <fcntl.h>
-#include <limits.h>
-#include <time.h>
-#include <math.h>
-#include <stdarg.h>
-#include <errno.h>
-#include <signal.h>
-#include <netdb.h>
-#include <locale.h>
-#include <dirent.h>
-#include <termios.h>
-#include <pthread.h>
-#include <semaphore.h>
-#include <sys/stat.h>
-#include <sys/time.h>
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <arpa/inet.h>
-
-#include <iostream>
-#include <string>
-#include <cstdlib>
-#include <cstring>
-#include <list>
-#include <vector>
-#include <deque>
-#include <algorithm>
-
-// 采用stl标准库的命名空间std
-using namespace std;
-
-#endif
-
