--- _ftp.cpp
@@ -1,265 +1,0 @@
-/****************************************************************************************/
-/*   程序名：_ftp.cpp，此程序是freecplus框架的ftp客户端工具的类的定义文件。             */
-/*   作者：C语言技术网(www.freecplus.net) 日期：20190525。                              */
-/*                                                                                      */
-/*   freecplus开源许可协议正文如下：                                                    */
-/*     版权所有 (c) 2008-2020，码农有道，C语言技术网（www.freecplus.net）。             */
-/*     特此向任何得到本软件副本或相关文档的人授权：被授权人有权使用、复制、修改、       */
-/*   合并、出版、发布、散布、再授权和/或贩售软件及软件的副本，及授予被供应人同等        */
-/*   权利，只需服从以下义务：                                                           */
-/*     在软件和软件的所有副本中都必须包含以上版权声明和本许可声明。                     */
-/*     如果违返，无论在本软件的使用上或其他买卖交易中是否涉及合同，都属于侵权行为，     */
-/*   我们保留追究的权利。                                                               */
-/*   freecplus开源许可协议正文结束。                                                    */
-/****************************************************************************************/
-
-#include "_ftp.h"
-
-Cftp::Cftp() {
-  m_ftpconn = 0;
-
-  initdata();
-
-  FtpInit();
-
-  m_connectfailed = false;
-  m_loginfailed = false;
-  m_optionfailed = false;
-}
-
-Cftp::~Cftp() {
-  logout();
-}
-
-void Cftp::initdata() {
-  m_size = 0;
-
-  memset(m_mtime, 0, sizeof(m_mtime));
-}
-
-bool Cftp::login(const char *host, const char *username, const char *password, const int imode) {
-  if (m_ftpconn != 0) {
-	FtpQuit(m_ftpconn);
-	m_ftpconn = 0;
-  }
-
-  m_connectfailed = m_loginfailed = m_optionfailed = false;
-
-  if (FtpConnect(host, &m_ftpconn) == false) {
-	m_connectfailed = true;
-	return false;
-  }
-
-  if (FtpLogin(username, password, m_ftpconn) == false) {
-	m_loginfailed = true;
-	return false;
-  }
-
-  if (FtpOptions(FTPLIB_CONNMODE, (long)imode, m_ftpconn) == false) {
-	m_optionfailed = true;
-	return false;
-  }
-
-  return true;
-}
-
-bool Cftp::logout() {
-  if (m_ftpconn == 0)
-	return false;
-
-  FtpQuit(m_ftpconn);
-
-  m_ftpconn = 0;
-
-  return true;
-}
-
-bool Cftp::get(const char *remotefilename, const char *localfilename, const bool bCheckMTime) {
-  if (m_ftpconn == 0)
-	return false;
-
-  // 创建本地文件目录
-  MKDIR(localfilename);
-
-  char strlocalfilenametmp[301];
-  memset(strlocalfilenametmp, 0, sizeof(strlocalfilenametmp));
-  snprintf(strlocalfilenametmp, 300, "%s.tmp", localfilename);
-
-  // 获取远程服务器的文件的时间
-  if (mtime(remotefilename) == false)
-	return false;
-
-  // 取文件
-  if (FtpGet(strlocalfilenametmp, remotefilename, FTPLIB_IMAGE, m_ftpconn) == false)
-	return false;
-
-  // 判断文件获取前和获取后的时间，如果时间不同，表示文件已改变，返回失败
-  if (bCheckMTime == true) {
-	char strmtime[21];
-	strcpy(strmtime, m_mtime);
-
-	if (mtime(remotefilename) == false)
-	  return false;
-
-	if (strcmp(m_mtime, strmtime) != 0)
-	  return false;
-  }
-
-  // 重置文件时间
-  UTime(strlocalfilenametmp, m_mtime);
-
-  // 改为正式的文件
-  if (rename(strlocalfilenametmp, localfilename) != 0)
-	return false;
-
-  // 获取文件的大小
-  m_size = FileSize(localfilename);
-
-  return true;
-}
-
-bool Cftp::mtime(const char *remotefilename) {
-  if (m_ftpconn == 0)
-	return false;
-
-  memset(m_mtime, 0, sizeof(m_mtime));
-
-  char strmtime[21];
-  memset(strmtime, 0, sizeof(strmtime));
-
-  if (FtpModDate(remotefilename, strmtime, 14, m_ftpconn) == false)
-	return false;
-
-  AddTime(strmtime, m_mtime, 0 + 8 * 60 * 60, "yyyymmddhh24miss");
-
-  return true;
-}
-
-bool Cftp::size(const char *remotefilename) {
-  if (m_ftpconn == 0)
-	return false;
-
-  m_size = 0;
-
-  if (FtpSize(remotefilename, &m_size, FTPLIB_IMAGE, m_ftpconn) == false)
-	return false;
-
-  return true;
-}
-
-bool Cftp::site(const char *command) {
-  if (m_ftpconn == 0)
-	return false;
-
-  if (FtpSite(command, m_ftpconn) == false)
-	return false;
-
-  return true;
-}
-
-bool Cftp::chdir(const char *remotedir) {
-  if (m_ftpconn == 0)
-	return false;
-
-  if (FtpChdir(remotedir, m_ftpconn) == false)
-	return false;
-
-  return true;
-}
-
-bool Cftp::mkdir(const char *remotedir) {
-  if (m_ftpconn == 0)
-	return false;
-
-  if (FtpMkdir(remotedir, m_ftpconn) == false)
-	return false;
-
-  return true;
-}
-
-bool Cftp::rmdir(const char *remotedir) {
-  if (m_ftpconn == 0)
-	return false;
-
-  if (FtpRmdir(remotedir, m_ftpconn) == false)
-	return false;
-
-  return true;
-}
-
-bool Cftp::dir(const char *remotedir, const char *listfilename) {
-  if (m_ftpconn == 0)
-	return false;
-
-  if (FtpDir(listfilename, remotedir, m_ftpconn) == false)
-	return false;
-
-  return true;
-}
-
-bool Cftp::nlist(const char *remotedir, const char *listfilename) {
-  if (m_ftpconn == 0)
-	return false;
-
-  // 创建本地文件目录
-  MKDIR(listfilename);
-
-  if (FtpNlst(listfilename, remotedir, m_ftpconn) == false)
-	return false;
-
-  return true;
-}
-
-bool Cftp::put(const char *localfilename, const char *remotefilename, const bool bCheckSize) {
-  if (m_ftpconn == 0)
-	return false;
-
-  char strremotefilenametmp[301];
-  memset(strremotefilenametmp, 0, sizeof(strremotefilenametmp));
-  snprintf(strremotefilenametmp, 300, "%s.tmp", remotefilename);
-
-  if (FtpPut(localfilename, strremotefilenametmp, FTPLIB_IMAGE, m_ftpconn) == false)
-	return false;
-
-  if (FtpRename(strremotefilenametmp, remotefilename, m_ftpconn) == false)
-	return false;
-
-  // 判断已上传的文件的大小与本地文件是否相同，确保上传成功。
-  if (bCheckSize == true) {
-	if (size(remotefilename) == false)
-	  return false;
-
-	if (m_size != FileSize(localfilename))
-	  return false;
-  }
-
-  return true;
-}
-
-bool Cftp::ftpdelete(const char *remotefilename) {
-  if (m_ftpconn == 0)
-	return false;
-
-  if (FtpDelete(remotefilename, m_ftpconn) == false)
-	return false;
-
-  return true;
-}
-
-bool Cftp::ftprename(const char *srcremotefilename, const char *dstremotefilename) {
-  if (m_ftpconn == 0)
-	return false;
-
-  if (FtpRename(srcremotefilename, dstremotefilename, m_ftpconn) == false)
-	return false;
-
-  return true;
-}
-
-char *Cftp::response() {
-  if (m_ftpconn == 0)
-	return 0;
-
-  return FtpLastResponse(m_ftpconn);
-}
-
