+++ db/CMakeLists.txt
@@ -1,0 +1,1 @@
+project(freecplus-db)
