+++ demo/CMakeLists.txt
@@ -1,0 +1,1 @@
+project(freecplus-demo)
