+++ freecplus/util/signal_util.h
@@ -1,0 +1,11 @@
+
+#pragma once
+
+
+
+////////////////////////////////////////////////////////////////////////////
+// 以下代码用于文件传输系统，不属于freecplus框架。
+
+// 关闭全部的信号和输入输出
+void CloseIOAndSignal();
+
