+++ freecplus/util/time_util.h
@@ -1,0 +1,49 @@
+#pragma once
+
+#include <ctime>
+
+///////////////////////////////////// /////////////////////////////////////
+/*
+  取操作系统的时间。
+  stime：用于存放获取到的时间字符串。
+  timetvl：时间的偏移量，单位：秒，0是缺省值，表示当前时间，30表示当前时间30秒之后的时间点，-30表示当前时间30秒之前的时间点。
+  fmt：输出时间的格式，fmt每部分的含义：yyyy-年份；mm-月份；dd-日期；hh24-小时；mi-分钟；ss-秒，
+  缺省是"yyyy-mm-dd hh24:mi:ss"，目前支持以下格式：
+  "yyyy-mm-dd hh24:mi:ss"
+  "yyyymmddhh24miss"
+  "yyyy-mm-dd"
+  "yyyymmdd"
+  "hh24:mi:ss"
+  "hh24miss"
+  "hh24:mi"
+  "hh24mi"
+  "hh24"
+  "mi"
+  注意：
+    1）小时的表示方法是hh24，不是hh，这么做的目的是为了保持与数据库的时间表示方法一致；
+    2）以上列出了常用的时间格式，如果不能满足你应用开发的需求，请修改源代码timetostr函数增加更多的格式支持；
+    3）调用函数的时候，如果fmt与上述格式都匹配，stime的内容将为空。
+    4）时间的年份是四位，其它的可能是一位和两位，如果不足两位，在前面补0。
+*/
+void LocalTime(char *stime, const char *fmt = 0, const int timetvl = 0);
+
+// 把整数表示的时间转换为字符串表示的时间。
+// ltime：整数表示的时间。
+// stime：字符串表示的时间。
+// fmt：输出字符串时间stime的格式，与LocalTime函数的fmt参数相同，如果fmt的格式不正确，stime将为空。
+void timetostr(const time_t ltime, char *stime, const char *fmt = 0);
+
+// 把字符串表示的时间转换为整数表示的时间。
+// stime：字符串表示的时间，格式不限，但一定要包括yyyymmddhh24miss，一个都不能少，顺序也不能变。
+// 返回值：整数表示的时间，如果stime的格式不正确，返回-1。
+time_t strtotime(const char *stime);
+
+// 把字符串表示的时间加上一个偏移的秒数后得到一个新的字符串表示的时间。
+// in_stime：输入的字符串格式的时间，格式不限，但一定要包括yyyymmddhh24miss，一个都不能少，顺序也不能变。
+// out_stime：输出的字符串格式的时间。
+// timetvl：需要偏移的秒数，正数往后偏移，负数往前偏移。
+// fmt：输出字符串时间out_stime的格式，与LocalTime函数的fmt参数相同。
+// 注意：in_stime和out_stime参数可以是同一个变量的地址，如果调用失败，out_stime的内容会清空。
+// 返回值：true-成功，false-失败，如果返回失败，可以认为是in_stime的格式不正确。
+bool AddTime(const char *in_stime, char *out_stime, const int timetvl, const char *fmt = 0);
+///////////////////////////////////// /////////////////////////////////////
