+++ ini/hssms.xml
@@ -1,20 +1,20 @@
 <?xml version="1.0" encoding="gbk" ?>
 <root>
-    <!-- ³ÌĞòÔËĞĞµÄÈÕÖ¾ÎÄ¼şÃû¡£ -->
+    <!-- ç¨‹åºè¿è¡Œçš„æ—¥å¿—æ–‡ä»¶åã€‚ -->
     <logpath>/log/hssms</logpath>
 
-    <!-- Êı¾İ¿âÁ¬½Ó²ÎÊı¡£ -->
+    <!-- æ•°æ®åº“è¿æ¥å‚æ•°ã€‚ -->
     <connstr>hssms/smspwd@hssmszx</connstr>
 
-    <!-- Êı¾İÎÄ¼ş´æ·ÅµÄ¸ùÄ¿Â¼¡£ -->
+    <!-- æ•°æ®æ–‡ä»¶å­˜æ”¾çš„æ ¹ç›®å½•ã€‚ -->
     <datapath>/data/hssms</datapath>
 
-    <!-- ÖĞĞÄ·şÎñÆ÷µÄipµØÖ·¡£ -->
+    <!-- ä¸­å¿ƒæœåŠ¡å™¨çš„ipåœ°å€ã€‚ -->
     <serverip>192.168.1.1</serverip>
 
-    <!-- ÖĞĞÄ·şÎñÆ÷µÄÍ¨ĞÅ¶Ë¿Ú¡£ -->
+    <!-- ä¸­å¿ƒæœåŠ¡å™¨çš„é€šä¿¡ç«¯å£ã€‚ -->
     <port>5058</port>
 
-    <!-- ÊÇ·ñ²ÉÓÃ³¤Á¬½Ó£¬true-ÊÇ£»false-·ñ¡£ -->
+    <!-- æ˜¯å¦é‡‡ç”¨é•¿è¿æ¥ï¼Œtrue-æ˜¯ï¼›false-å¦ã€‚ -->
     <online>true</online>
 </root>
